Initial Commit of Dijkstra Visualization App
Create Class for Moveable Circles that will serve as the nodes in our
graph.  Future installments of the base project (creating and
visualizing graphs) will include the options for other shapes.  As such
I plan to extract much of the logic from this class into a MoveableShape
abstract class

Add all of the logic necessary for displaying the shapes and whatnot on
the screen. Once logic becomes more complex (when we actually need to
do dijkstra, etc) the display logic may need some tinkering so that we
have a good balance between display logic and business logic.  We will
also make some changes to how the multithreading is being done.

Add logic for selecting multiple nodes and connecting them.  Distance
between the nodes appears.  This should be modified so that the distance
is being shrunk by some factor.  May also add a toggle for showing vs
not showing the distances on the edges.

To Do in Next Commit:
---------------------
1)Add ability to add more nodes in.
 i)ensure that this does not interfere with the ability to create edges
 ii)Also add logic so that we cannot have multiple edges between the
    same nodes

2)Make the edges update as we move our nodes with edges attached.  This
will take some modification to the current data structures and classes.

3)Toggle the selection as soon as two nodes have been selected and the
edge has been created.  We could add an option page so that the user can
choose whether or not they want to manually make all of the toggles.